add new functionality: a task can only be deleted by his author

+++ src/AppBundle/Controller/TaskController.php
@@ -85,11 +85,15 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        if ($task->getUser() === $this->getUser()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        } else {
+            $this->addFlash('error', 'Vous n\'êtes pas l\'auteur de cette tâche.');
+        }
 
         return $this->redirectToRoute('task_list');
     }
